Added admin service and updated dao

+++ src/main/java/com/mitsko/mrdb/dao/AdminDAO.java
@@ -4,5 +4,5 @@ import com.mitsko.mrdb.entity.util.Status;
 
 public interface AdminDAO {
     void refreshStatus(Status newStatus, String login);
-    void refreshAverageRating(float newRating, String login);
+    void refreshAverageRating(int newRating, String login);
 }
